HW2: переработка задачи с файлами
