+++ homeworks/homework_02/table/file_encodings.py
@@ -1,5 +1,5 @@
 """
-Модуль содержащий методы для определения кодировки файла
+Методы для определения кодировки файла
 
 альтернативные варианты:
     import chardet
